Split Daytona sandbox large files
